Refactor output module headers to impl files
